requirements.txt
gmic==2.9.2
pillow==8.1.0
